requirements.txt
Flask==2.0.1

Flask-Login==0.5.0

Flask-MongoEngine==1.0.0

Werkzeug==2.0.1

python-dateutil==2.8.2

pymongo==3.12.0

pytz==2023.3

Pillow==9.5.0

email-validator==2.0.0

python-dotenv==1.0.0

bcrypt==4.0.1

Flask-Mail==0.9.1

Flask-WTF==1.1.1

requests==2.31.0

oauthlib==3.2.2

PyJWT==2.8.0
